backend/requirements.txt
# FastAPI and Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Database
sqlalchemy>=2.0.36
aiosqlite>=0.20.0

# Document Processing
python-pptx>=1.0.2
pypdfium2>=4.30.0
python-docx>=1.1.2
Pillow>=11.0.0

# OCR
pytesseract>=0.3.13

# PDF Generation
weasyprint>=63.0

# AI/LLM - OpenRouter (OpenAI-compatible)
openai>=1.50.0

# Utilities
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
httpx>=0.28.0
aiofiles>=24.1.0

# GitHub Integration
PyGithub>=2.5.0

# Markdown processing
markdown>=3.7
